requirements.txt
fastapi
uvicorn
pdfplumber
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
